feat: linux windows 支持 kiosk 模式
