+++ packages/electron-screenshots/src/getDisplay.ts
@@ -1,0 +1,19 @@
+import { Rectangle, screen } from 'electron'
+
+export interface Display extends Rectangle {
+  id: number
+}
+
+export default (): Display => {
+  const point = screen.getCursorScreenPoint()
+  const { id, bounds } = screen.getDisplayNearestPoint(point)
+
+  // https://github.com/nashaofu/screenshots/issues/98
+  return {
+    id,
+    x: Math.floor(bounds.x),
+    y: Math.floor(bounds.y),
+    width: Math.floor(bounds.width),
+    height: Math.floor(bounds.height)
+  }
+}
